backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
requests
pandas
opencv-python
scikit-learn
matplotlib